--- src/app/account/account.module.ts
@@ -1,15 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-import { AccountRoutingModule } from './account-routing.module';
-import { AuthModule } from './auth/auth.module';
-
-@NgModule({
-  declarations: [],
-  imports: [
-    CommonModule,
-    AccountRoutingModule,
-    AuthModule
-  ]
-})
-export class AccountModule { }
