--- src/app/account/auth/passwordreset/passwordreset.component.html
@@ -1,64 +1,0 @@
-<div class="account-pages my-5 pt-sm-5">
-  <div class="container">
-    <div class="row justify-content-center">
-      <div class="col-md-8 col-lg-6 col-xl-5">
-        <div class="card overflow-hidden">
-          <div class="bg-soft bg-primary">
-            <div class="row">
-              <div class="col-7">
-                <div class="text-primary p-4">
-                  <h5 class="text-primary"> Reset Password</h5>
-                  <p>Re-Password with Skote.</p>
-                </div>
-              </div>
-              <div class="col-5 align-self-end">
-                <img src="assets/images/profile-img.png" alt="" class="img-fluid">
-              </div>
-            </div>
-          </div>
-          <div class="card-body pt-0">
-            <div>
-              <a routerLink="/">
-                <div class="avatar-md profile-user-wid mb-4">
-                  <span class="avatar-title rounded-circle bg-light">
-                    <img src="assets/images/logo.svg" alt="" class="rounded-circle" height="34">
-                  </span>
-                </div>
-              </a>
-            </div>
-
-            <form class="needs-validation" name="resetForm" [formGroup]="resetForm" (ngSubmit)="onSubmit()" novalidate>
-
-              <ngb-alert type="danger" *ngIf="error" [dismissible]="false">{{ error }}</ngb-alert>
-
-              <div class="mb-3">
-                <label for="email">Email</label>
-
-                <input type="email" formControlName="email" class="form-control"
-                  [ngClass]="{ 'is-invalid': submitted && f.email.errors }" id="email" placeholder="Email" />
-
-                <div *ngIf="submitted && f.email.errors" class="invalid-feedback">
-                  <div *ngIf="f.email.errors.required">Email is required</div>
-                  <div *ngIf="f.email.errors.email">Email must be a valid email address</div>
-                </div>
-              </div>
-              <div class="text-end">
-                <button class="btn btn-primary w-md" type="submit">Reset</button>
-              </div>
-            </form>
-
-          </div> <!-- end card-body-->
-        </div>
-        <!-- end card -->
-        <div class="mt-5 text-center">
-          <p>Remember It ? <a routerLink="/account/login" class="fw-medium text-primary"> Sign In here</a> </p>
-          <p>© {{year}} Skote. Crafted with <i class="mdi mdi-heart text-danger"></i> by Themesbrand</p>
-        </div>
-
-      </div> <!-- end col -->
-    </div>
-    <!-- end row -->
-  </div>
-  <!-- end container -->
-</div>
-<!-- end page -->
