+++ src/app/app-routing.module.ts
@@ -1,13 +1,12 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-
-import { AuthGuard } from './core/guards/auth.guard';
 import { LayoutComponent } from './layouts/layout.component';
 import { Page404Component } from './extrapages/page404/page404.component';
+import { AuthGuard } from './core/guards/auth.guard';
 
 const routes: Routes = [
-  { path: 'account', loadChildren: () => import('./account/account.module').then(m => m.AccountModule) },
-  // tslint:disable-next-line: max-line-length
+ { path: 'usuario', loadChildren: () => import('./usuario/usuario.module').then(m => m.UsuarioModule) },
+ { path: '', redirectTo: 'login', pathMatch: 'full' }, // Redirige a VistaGeneral por defecto
   { path: '', component: LayoutComponent, loadChildren: () => import('./pages/pages.module').then(m => m.PagesModule), canActivate: [AuthGuard] },
   { path: 'pages', loadChildren: () => import('./extrapages/extrapages.module').then(m => m.ExtrapagesModule), canActivate: [AuthGuard] },
   { path: '**', component: Page404Component },
