+++ src/app/layouts/sidebar/sidebar.component.html
@@ -10,14 +10,6 @@
       <ng-container *ngFor="let item of menuItems">
         <li class="menu-title" *ngIf="item.isTitle">{{ item.label | translate }}</li>
         <li *ngIf="!item.isTitle && !item.isLayout">
-          <a *ngIf="hasItems(item)" href="javascript:void(0);" class="is-parent"
-            [ngClass]="{'has-arrow': !item.badge, 'has-dropdown': item.badge}">
-            <i class="bx {{item.icon}}" *ngIf="item.icon"></i>
-            <span> {{ item.label | translate }}</span>
-            <span class="badge rounded-pill float-end bg-{{item.badge.variant}}" *ngIf="item.badge">{{item.badge.text |
-              translate}}</span>
-          </a>
-
           <a [routerLink]="item.link" *ngIf="!hasItems(item)" class="side-nav-link-ref" routerLinkActive="active">
             <i class="bx {{ item.icon }}" *ngIf="item.icon"></i>
             <span> {{ item.label | translate }}</span>
