+++ src/app/usuario/login/login.component.html
@@ -1,0 +1,58 @@
+<div class="account-pages my-5 pt-sm-5">
+  <div class="container">
+    <div class="card-body">
+      <div class="row justify-content-center">
+        <div class="col-md-8 col-lg-6 col-xl-5">
+          <div class="card overflow-hidden">
+              <div class="row mt-3">
+                <div class="col-12 center-self-end">
+                  <img src="assets/images/minerva2.png" alt="" class="img-fluid smaller-image">
+                </div>
+              </div>
+      
+            <div class="card-body pt-0">
+              <div class="p-2">
+                <form class="form-horizontal">
+                  <div class="mb-3 input-group"> <!-- Agregamos la clase "input-group" para agrupar el icono y el input -->
+                    <div class="input-group-prepend"> <!-- Agregamos un div para el icono -->
+                      <span class="input-group-text"><i class="mdi mdi-lock mr-1 "></i></span>
+                    </div>
+                    <input type="email" formControlName="email" class="form-control" id="email" placeholder="Email" [ngClass]="" />
+                    <div *ngIf="" class="invalid-feedback">
+                      <div *ngIf="">Usuario requerido</div>
+                      <div *ngIf="">Usuario incorrecto</div>
+                    </div>
+                  </div>
+                
+                  <div class="mb-3 input-group"> <!-- Repetimos el mismo proceso para el segundo input -->
+                    <div class="input-group-prepend">
+                      <span class="input-group-text"><i class="mdi mdi-lock mr-1"></i></span>
+                    </div>
+                    <input type="password" class="form-control" formControlName="password" placeholder="Enter password" [ngClass]=""
+                      aria-label="Password" aria-describedby="password-addon">
+                    <div *ngIf="" class="invalid-feedback">
+                      <span *ngIf="">Contraseña requerida</span>
+                    </div>
+                    <div class="input-group-append"> <!-- Agregamos un div para el botón de ojo -->
+                      <button class="btn btn-light" type="button" id="password-addon"><i class="mdi mdi-eye-outline"></i></button>
+                    </div>
+                  </div>
+                  <div class="mt-3 d-grid">
+                    <button class="btn btn-primary" type="submit">Ingresar</button>
+                  </div>
+                  <div class="mt-4 text-center">
+                    <a routerLink="/usuario/reset-password" class="text-muted"><i class="mdi mdi-lock mr-1"></i> ¿Olvidó su contraseña?</a>
+                  </div>
+                </form>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+   
+  </div>
+
+  <!-- end container -->
+</div>
+<!-- end page -->
