+++ src/app/usuario/login/login.component.scss
@@ -1,0 +1,11 @@
+.smaller-image {
+    max-width: 25%;
+    display: block; 
+    margin: 0 auto; 
+   
+}
+
+
+
+
+
