+++ src/app/usuario/usuario.module.ts
@@ -1,0 +1,18 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { UsuarioRoutingModule } from './usuario-routing.module';
+import { LoginComponent } from './login/login.component';
+
+
+@NgModule({
+  declarations: [
+    LoginComponent
+  ],
+  imports: [
+    CommonModule,
+    UsuarioRoutingModule,
+   
+  
+  ]
+})
+export class UsuarioModule { }
